# server/swagger.yaml
openapi: 3.0.3
info:
  title: EventKnit API
  description: |
    EventKnit is a comprehensive ticketing platform API that enables event management, ticket sales, payments, and real-time communications.

    ## Authentication
    Most endpoints require JWT authentication. Include the access token in the Authorization header:
    ```
    Authorization: Bearer <your_access_token>
    ```

    ## Base URL
    All API routes are prefixed with `/api/v1`

    ## Rate Limiting
    API requests are rate-limited to prevent abuse. Rate limits vary by endpoint type.
  version: 1.0.0
  contact:
    name: EventKnit Support
    email: support@eventknit.com
  license:
    name: ISC
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.eventknit.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login/registration
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Registration:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        ticketTypeId:
          type: string
          format: uuid
        quantity:
          type: integer
        totalAmount:
          type: number
          format: decimal
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, CHECKED_IN]
        attendeeInfo:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
            phoneNumber:
              type: string
        paymentStatus:
          type: string
          enum: [PENDING, COMPLETED, FAILED, REFUNDED]
        qrCode:
          type: string
        createdAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Unauthorized access
...
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Resource not found
...
paths:
...
  /tickets/{registrationId}:
    get:
      tags:
        - Tickets
      summary: Get ticket by registration ID
      description: Retrieve ticket information for a specific registration
      operationId: getTicket
      security:
        - BearerAuth: []
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Registration'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tickets/{registrationId}/view:
    get:
      tags:
        - Tickets
      summary: Get ticket publicly
      description: View ticket information with email verification (public endpoint)
      operationId: getTicketPublic
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Ticket retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Registration'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tickets/{registrationId}/download:
    get:
      tags:
        - Tickets
      summary: Download ticket as PDF
      description: Download ticket in PDF format
      operationId: downloadTicketPDF
      security:
        - BearerAuth: []
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: PDF generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tickets/{registrationId}/resend:
    post:
      tags:
        - Tickets
      summary: Resend ticket email
      description: Resend ticket confirmation email (rate limited to 3 per hour)
      operationId: resendTicketEmail
      security:
        - BearerAuth: []
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== TICKET TRANSFERS ====================
...
security:
  - BearerAuth: []